Move: redux 설정 관련 코드들을 한 디렉토리에 묶어서 정리
