+++ client/src/app/redux/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import rootReducer from "@app/rootReducer";
+import rootReducer from "./rootReducer";
 
 export const store = configureStore({
   reducer: rootReducer
