+++ client/src/component/photo-list/content/SearchInput.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useCallback } from 'react';
-import { useAppSelector, useAppDispatch } from '@app/reduxHooks';
+import { useAppSelector, useAppDispatch } from '@app/redux/reduxHooks';
 import { addLabel } from './searchSlice';
 import SearchBar from '@component/searchbar/SearchBar';
 import CardHeader from '@component/card/basic/CardHeader';
