+++ client/src/page/admin/voucher/list/content/FilterCheck.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useCallback } from 'react';
 import { useQuery } from 'react-query';
 import { usePopper } from 'react-popper';
-import { useAppSelector, useAppDispatch } from '@app/reduxHooks';
+import { useAppSelector, useAppDispatch } from '@app/redux/reduxHooks';
 import { AxiosError, AxiosResponse } from 'axios';
 import { ErrorType } from '@util/commonAPI';
 import * as queryKey from '@util/queryKey';
