+++ client/src/page/admin/voucher/list/content/SearchLabelList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAppSelector, useAppDispatch } from '@app/reduxHooks';
+import { useAppSelector, useAppDispatch } from '@app/redux/reduxHooks';
 import CardHeader from '@component/card/basic/CardHeader';
 import SearchLabel from '@component/label/SearchLabel';
 import { VoucherStateName, toggleGroup, toggleMember, changeVoucherFilter, removeName } from '../voucherListSlice';
